CLN: remove PyFibreWorkflow
