+++ pyfibre/api.py
@@ -16,6 +16,3 @@ from pyfibre.model.multi_image.shg_pl_trans_image import SHGPLTransImage
 from pyfibre.model.tools.bd_cluster import BDFilter
 from pyfibre.model.tools.fibre_assigner import FibreAssigner
 from pyfibre.model.tools.fire_algorithm import FIREAlgorithm
-
-from pyfibre.model.pyfibre_runner import PyFibreRunner
-from pyfibre.model.pyfibre_workflow import PyFibreWorkflow
