+++ pyfibre/io/pyfibre_workflow_io.py
@@ -1,10 +1,10 @@
 from pyfibre.io.utilities import load_json
 
-from pyfibre.model.pyfibre_workflow import PyFibreWorkflow
+from pyfibre.model.pyfibre_runner import PyFibreRunner
 
 
 def load_pyfibre_workflow(filepath):
 
     data = load_json(filepath)
 
-    return PyFibreWorkflow(**data)
+    return PyFibreRunner(**data)
